Refactored logger logic (#1120)
* Refactored logger logic


* Cleaned up transport.js


* Re-order


* comments
