+++ financial-templates-lib/logger/RoboCaller.js
@@ -1,24 +1,18 @@
-require("dotenv").config();
-
 class RoboCaller {
-  constructor() {
+  constructor(params) {
     if (
-      !process.env.TWILIO_SID || // require the account SID
-      !process.env.TWILIO_AUTH || // require the account authentication key
-      !process.env.DRI_NUMBER1 || // require at least 1 number to call
-      !process.env.TWILIO_FROM_NUMBER // require the number to originate the call from
+      !params.twilioSid || // require the account SID
+      !params.twilioAuth || // require the account authentication key
+      !params.twilioFrom || // require the number to originate the call from
+      !params.twilioCallNumbers // require at least 1 number to call
     ) {
       console.error("Missing config variable. RoboCaller Disabled");
       return;
     }
-    this.client = require("twilio")(process.env.TWILIO_SID, process.env.TWILIO_AUTH);
+    this.client = require("twilio")(params.twilioSid, params.twilioAuth);
 
-    this.numbersToCall = [];
-    for (const envVariable in process.env) {
-      if (envVariable.startsWith("DRI_NUMBER")) {
-        this.numbersToCall.push(process.env[envVariable]);
-      }
-    }
+    this.twilioFrom = params.twilioFrom;
+    this.twilioCallNumbers = params.twilioCallNumbers;
   }
   canPlaceCall = () => {
     return this.client != undefined;
@@ -29,12 +23,12 @@ class RoboCaller {
       return;
     }
 
-    for (const number of this.numbersToCall) {
+    for (const number of this.twilioCallNumbers) {
       try {
         const callResponse = await this.client.calls.create({
           twiml: this.generateTwiML(message),
           to: number,
-          from: process.env.TWILIO_FROM_NUMBER
+          from: this.twilioFrom
         });
       } catch (error) {
         console.error("something went wrong", error);
