+++ liquidator/liquidator.js
@@ -1,7 +1,7 @@
 // When running this script it assumed that the account has enough tokens and allowance from the unlocked truffle
 // wallet to run the liquidations. Future versions will deal with generating additional synthetic tokens from EMPs as the bot needs.
 
-const { Logger } = require("../financial-templates-lib/Logger");
+const { Logger } = require("../financial-templates-lib/logger/Logger");
 
 class Liquidator {
   constructor(expiringMultiPartyClient, gasEstimator, account) {
